overworked merge conflicts

+++ WindowsFormsApp1/View/Bookadministration.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Data;
 using System.Data.SqlClient;
+using System.Drawing;
 using System.Text;
 using System.Windows.Forms;
 using WindowsFormsApp1.Controller;
+using WindowsFormsApp1.Model;
 
 namespace WindowsFormsApp1
 {
@@ -39,11 +41,18 @@ namespace WindowsFormsApp1
             dataAdapter = new SqlDataAdapter("SELECT * FROM Books", connectionString);
             dataAdapter.Fill(table);
             addButton("Reservieren");
-            addButton("Ausleihe");
-            addButton("Rückgabe");
-
-            addButton("Bearbeiten");
-
+            if (CurrentUser.getAdmin())
+            {
+                addButton("Ausleihe");
+                addButton("Rückgabe");
+                addButton("Löschen");
+                addButton("Bearbeiten");
+            } else
+            {
+                btnAdd.Hide();
+                button2.Hide();
+                label1.Text = "Alle Bücher";
+            }
             table.Columns.Add("_RowString", typeof(string));
             dataGridView1.Columns["_RowString"].Visible = false;
         }
@@ -62,7 +71,10 @@ namespace WindowsFormsApp1
             try
             {
                 table.Clear();
-                dataAdapter = new SqlDataAdapter("SELECT * FROM Books", connectionString);
+                string query;
+                if (CurrentUser.getAdmin()) query = "select * from Books";
+                else query = "select * from Books b Where NOT b.ISBN = ANY (SELECT FK_ISBN FROM Issues Where IssueState ='aus' AND FK_ISBN =b.ISBN)";
+                dataAdapter = new SqlDataAdapter(query, connectionString);
                 SqlCommandBuilder commandBuilder = new SqlCommandBuilder(dataAdapter);
                 dataAdapter.Fill(table);
                 int z = 0;
@@ -70,7 +82,6 @@ namespace WindowsFormsApp1
                 {
                    
                     if (BookAdminModel.IsLend(dataRow[0].ToString(), dataRow[1].ToString())) {
-                           
                         DataGridViewRow row = dataGridView1.Rows[z];
                         row.DefaultCellStyle.BackColor = Color.Blue;
                     }
@@ -83,56 +94,14 @@ namespace WindowsFormsApp1
                     dataRow["_RowString"] = sb.ToString();
                     z++;
                 }
-
-                // Hide filter string column
-                dataGridView1.Columns["_RowString"].Visible = false;
-                dataGridView1.AutoResizeColumns(
-                    DataGridViewAutoSizeColumnsMode.AllCells);
-
-                dataAdapter.Update((DataTable)bindingSource1.DataSource);
-               
-
-              
-
+                // Resize the DataGridView columns to fit the newly loaded content.
+                dataGridView1.AutoResizeColumns(DataGridViewAutoSizeColumnsMode.AllCells);
             }
             catch (SqlException sql)
             {
                 MessageBox.Show("Keine Einträge vorhanden! " + sql.Message);
             }
         }
-
-        private void addButton(string buttonName)
-        {
-            DataGridViewButtonColumn button = new DataGridViewButtonColumn();
-            button.Name = buttonName;
-            button.Text = buttonName;
-            if (dataGridView1.Columns[buttonName] == null)
-            {
-                dataGridView1.Columns.Add(button);
-            }
-        }
-
-        private void Form1_Load(object sender, EventArgs e)
-        {
-            dataGridView1.DataSource = bindingSource1;
-            GetData("select * from Books");
-            if (!CurrentUser.getAdmin())
-            {
-                label1.Text = "Alle Bücher";
-
-                dataGridView1.Columns["Ausleihe"].Visible = false;
-                dataGridView1.Columns["Rückgabe"].Visible = false;
-                dataGridView1.Columns["Bearbeiten"].Visible = false;
-                dataGridView1.Columns["Löschen"].Visible = false;
-                btnAdd.Visible = false;
-                button2.Visible = false;
-
-            }
-            else { button2.Visible = false; }
-            dataAdapter.Update((DataTable)bindingSource1.DataSource);
-
-        }
-
         private int HeaderPosition(string value)
         {
             return dataGridView1.Columns[value].Index;
@@ -141,25 +110,20 @@ namespace WindowsFormsApp1
         {
             string inventoryNumber = dataGridView1.Rows[e.RowIndex].Cells[HeaderPosition("Inventar_Number")].Value.ToString();
             string iSBN = dataGridView1.Rows[e.RowIndex].Cells[HeaderPosition("ISBN")].Value.ToString();
-
             if (e.ColumnIndex == HeaderPosition("Reservieren"))
             {
                 Controller.ReserveBook(iSBN, inventoryNumber);
             }
-
            
-            if (e.ColumnIndex == HeaderPosition("Rückgabe"))
-
-            {
-                Controller.ReturnBook(iSBN, inventoryNumber);
-            }
-
-            if (e.ColumnIndex == HeaderPosition("Ausleihe"))
-
-            {
+            if (CurrentUser.getAdmin()) { 
+                if (e.ColumnIndex == HeaderPosition("Rückgabe"))
+                {
+                    Controller.ReturnBook(iSBN, inventoryNumber);
+                }
+                if (e.ColumnIndex == HeaderPosition("Ausleihe"))
+                {
                 Controller.LendBook(iSBN, inventoryNumber);
-            }
-
+                }
                 if (e.ColumnIndex == HeaderPosition("Löschen"))
                 {
                     Controller.Delete(iSBN, inventoryNumber);
@@ -168,8 +132,7 @@ namespace WindowsFormsApp1
                 {
                     Controller.ShowBookInformation(iSBN, inventoryNumber);
                 }
-
-            
+            }
             RefreshData();
         }
         private void button2_Click(object sender, EventArgs e)
@@ -187,35 +150,13 @@ namespace WindowsFormsApp1
         }
         public void RefreshData()
         {
-
-            dataGridView1.DataSource = bindingSource1;
-            if (!CurrentUser.getAdmin()) { 
-                label1.Text ="Verfügbare Bücher";
-            
-            GetData("select * from Books b Where NOT b.ISBN = ANY (SELECT FK_ISBN FROM Issues Where IssueState ='aus' AND FK_ISBN =b.ISBN)");
-            if(textBox1.Text != "Suche") table.DefaultView.RowFilter = string.Format("[_RowString] LIKE '%{0}%'", textBox1.Text);
-           
-                dataGridView1.Columns["Ausleihe"].Visible = false;
-                dataGridView1.Columns["Rückgabe"].Visible = false;
-                dataGridView1.Columns["Bearbeiten"].Visible = false;
-                dataGridView1.Columns["Löschen"].Visible = false;
-                btnAdd.Visible = false;
-                button2.Visible = false;
-
-            }
-            else { button1.Visible = false;
-
-                GetData("select * from Books");}
-    
-                dataAdapter.Update((DataTable)bindingSource1.DataSource);
-
-
+            GetData();
         }
        
         public void GetAllData()
         {
             dataGridView1.DataSource = bindingSource1;
-            GetData("select * from Books");
+            // GetData("select * from Books");
             if (textBox1.Text != "Suche") table.DefaultView.RowFilter = string.Format("[_RowString] LIKE '%{0}%'", textBox1.Text);
             if (!CurrentUser.getAdmin())
             {
